fixed up SelectProvider
