+++ src/fiatExchanges/SelectProvider.tsx
@@ -1,56 +1,45 @@
+import { RouteProp } from '@react-navigation/native'
+import { StackScreenProps } from '@react-navigation/stack'
 import React, { useState } from 'react'
-import {
-  LayoutAnimation,
-  StyleSheet,
-  Text,
-  View,
-  Image,
-  ActivityIndicator,
-  ScrollView,
-} from 'react-native'
-import Touchable from 'src/components/Touchable'
+import { useAsync } from 'react-async-hook'
+import { useTranslation } from 'react-i18next'
+import { ActivityIndicator, ScrollView, StyleSheet, Text, View } from 'react-native'
+import { useDispatch, useSelector } from 'react-redux'
+import { showError } from 'src/alert/actions'
+import { FiatExchangeEvents } from 'src/analytics/Events'
+import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
+import { ErrorMessages } from 'src/app/ErrorMessages'
 import BackButton from 'src/components/BackButton'
-import Expandable from 'src/components/Expandable'
+import Dialog from 'src/components/Dialog'
+import Touchable from 'src/components/Touchable'
+import { PaymentMethodSection } from 'src/fiatExchanges/PaymentMethodSection'
+import i18n from 'src/i18n'
+import { getLocalCurrencyCode } from 'src/localCurrency/selectors'
 import { emptyHeader } from 'src/navigator/Headers'
+import { navigate } from 'src/navigator/NavigationService'
+import { Screens } from 'src/navigator/Screens'
+import { StackParamList } from 'src/navigator/types'
+import { userLocationDataSelector } from 'src/networkInfo/selectors'
 import colors from 'src/styles/colors'
 import fontStyles from 'src/styles/fonts'
+import variables from 'src/styles/variables'
+import { CiCoCurrency, Currency } from 'src/utils/currencies'
+import { navigateToURI } from 'src/utils/linking'
+import Logger from 'src/utils/Logger'
+import { currentAccountSelector } from 'src/web3/selectors'
 import {
+  CICOFlow,
+  CicoQuote,
+  fetchLegacyMobileMoneyProviders,
   fetchProviders,
+  filterLegacyMobileMoneyProviders,
+  getQuotes,
+  isSimplexQuote,
+  LegacyMobileMoneyProvider,
+  PaymentMethod,
   ProviderQuote,
   sortQuotes,
-  isSimplexQuote,
-  getQuotes,
-  CicoQuote,
-  getFeeValueFromQuotes,
-  SimplexQuote,
-  PaymentMethod,
-  fetchLocalCicoProviders,
-  getAvailableLocalProviders,
-  LocalCicoProvider,
-  CICOFlow,
 } from './utils'
-import { StackScreenProps } from '@react-navigation/stack'
-import { StackParamList } from 'src/navigator/types'
-import { Screens } from 'src/navigator/Screens'
-import { useDispatch, useSelector } from 'react-redux'
-import { userLocationDataSelector } from 'src/networkInfo/selectors'
-import { currentAccountSelector } from 'src/web3/selectors'
-import { getLocalCurrencyCode } from 'src/localCurrency/selectors'
-import { CiCoCurrency, Currency } from 'src/utils/currencies'
-import { useAsync } from 'react-async-hook'
-import Logger from 'src/utils/Logger'
-import { showError } from 'src/alert/actions'
-import variables from 'src/styles/variables'
-import { ErrorMessages } from 'src/app/ErrorMessages'
-import CurrencyDisplay from 'src/components/CurrencyDisplay'
-import { useTranslation } from 'react-i18next'
-import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
-import { FiatExchangeEvents } from 'src/analytics/Events'
-import { navigate } from 'src/navigator/NavigationService'
-import { navigateToURI } from 'src/utils/linking'
-import Dialog from 'src/components/Dialog'
-import i18n from 'src/i18n'
-import { RouteProp } from '@react-navigation/native'
 
 const TAG = 'SelectProviderScreen'
 
@@ -64,7 +53,7 @@ export default function SelectProviderScreen({ route, navigation }: Props) {
   const [noPaymentMethods, setNoPaymentMethods] = useState(false)
   const { flow } = route.params
 
-  const currencyToBuy = {
+  const digitalAsset = {
     [Currency.Celo]: CiCoCurrency.CELO,
     [Currency.Dollar]: CiCoCurrency.CUSD,
     [Currency.Euro]: CiCoCurrency.CEUR,
@@ -80,21 +69,21 @@ export default function SelectProviderScreen({ route, navigation }: Props) {
         userLocation,
         walletAddress: account,
         fiatCurrency: localCurrency,
-        digitalAsset: currencyToBuy,
+        digitalAsset,
         fiatAmount: route.params.amount.fiat,
         digitalAssetAmount: route.params.amount.crypto,
         txType: flow === CICOFlow.CashIn ? 'buy' : 'sell',
       })
 
-      const localProviders = await fetchLocalCicoProviders()
-      const availableLocalProviders = getAvailableLocalProviders(
-        localProviders,
+      const rawLegacyMobileMoneyProviders = await fetchLegacyMobileMoneyProviders()
+      const legacyMobileMoneyProviders = filterLegacyMobileMoneyProviders(
+        rawLegacyMobileMoneyProviders,
         flow,
         userLocation.countryCodeAlpha2,
-        currencyToBuy
+        digitalAsset
       )
 
-      return { providers, availableLocalProviders }
+      return { providers, legacyMobileMoneyProviders }
     } catch (error) {
       dispatch(showError(ErrorMessages.PROVIDER_FETCH_FAILED))
     }
@@ -144,9 +133,9 @@ export default function SelectProviderScreen({ route, navigation }: Props) {
         setNoPaymentMethods={setNoPaymentMethods}
         quoteOnPress={quoteOnPress}
       />
-      <LocalProviderMobileMoneySection
-        localProvider={asyncProviders.result?.availableLocalProviders[0]}
-        digitalAsset={currencyToBuy}
+      <LegacyMobileMoneySection
+        provider={asyncProviders.result?.legacyMobileMoneyProviders[0]}
+        digitalAsset={digitalAsset}
       />
       <ExchangesSection selectedCurrency={route.params.selectedCrypto} flow={flow} />
       <LimitedPaymentMethods visible={noPaymentMethods} />
@@ -222,23 +211,23 @@ function ExchangesSection({
   )
 }
 
-function LocalProviderMobileMoneySection({
-  localProvider,
+function LegacyMobileMoneySection({
+  provider,
   digitalAsset,
 }: {
-  localProvider: LocalCicoProvider | undefined
+  provider: LegacyMobileMoneyProvider | undefined
   digitalAsset: CiCoCurrency
 }) {
   const { t } = useTranslation()
-  if (!localProvider) {
+  if (!provider) {
     return null
   }
-  const goToLocalProvider = () => {
-    navigateToURI(localProvider[digitalAsset === CiCoCurrency.CUSD ? 'cusd' : 'celo'].url)
+  const goToProviderSite = () => {
+    navigateToURI(provider[digitalAsset === CiCoCurrency.CUSD ? 'cusd' : 'celo'].url)
   }
   return (
     <View style={styles.container}>
-      <Touchable onPress={goToLocalProvider}>
+      <Touchable onPress={goToProviderSite}>
         <View style={{ ...styles.expandableContainer, paddingVertical: 27 }}>
           <View style={styles.left}>
             <Text style={styles.category}>{t('selectProviderScreen.mobileMoney')}</Text>
@@ -246,171 +235,10 @@ function LocalProviderMobileMoneySection({
           </View>
 
           <View style={styles.right}>
-            <Text style={styles.linkToOtherScreen}>{localProvider.name}</Text>
+            <Text style={styles.linkToOtherScreen}>{provider.name}</Text>
           </View>
         </View>
       </Touchable>
-    </View>
-  )
-}
-
-interface PaymentMethodSectionProps {
-  paymentMethod: PaymentMethod
-  cicoQuotes: CicoQuote[]
-  setNoPaymentMethods: React.Dispatch<React.SetStateAction<boolean>>
-  quoteOnPress: (cicoQuote: CicoQuote) => void
-}
-
-export function PaymentMethodSection({
-  paymentMethod,
-  cicoQuotes,
-  setNoPaymentMethods,
-  quoteOnPress,
-}: PaymentMethodSectionProps) {
-  const { t } = useTranslation()
-  const sectionQuotes = cicoQuotes.filter(({ quote }) => quote.paymentMethod === paymentMethod)
-  const localCurrency = useSelector(getLocalCurrencyCode)
-
-  const isExpandable = sectionQuotes.length > 1
-  const [expanded, setExpanded] = useState(false)
-
-  const toggleExpanded = () => {
-    LayoutAnimation.easeInEaseOut()
-    setExpanded(!expanded)
-  }
-  if (!sectionQuotes.length) {
-    setNoPaymentMethods(true)
-    return null
-  }
-  const renderFeeAmount = (quote: SimplexQuote | ProviderQuote, postFix: string) => {
-    const feeAmount = getFeeValueFromQuotes(quote)
-
-    if (feeAmount === undefined) {
-      return null
-    }
-
-    return (
-      <Text>
-        <CurrencyDisplay
-          amount={{
-            value: 0,
-            localAmount: {
-              value: feeAmount,
-              currencyCode: localCurrency,
-              exchangeRate: 1,
-            },
-            currencyCode: localCurrency,
-          }}
-          showLocalAmount={true}
-          hideSign={true}
-          style={styles.fee}
-        />{' '}
-        {postFix}
-      </Text>
-    )
-  }
-  return (
-    <View style={styles.container}>
-      <Touchable
-        onPress={
-          isExpandable
-            ? toggleExpanded
-            : ((quoteOnPress(sectionQuotes[0]) as unknown) as () => void)
-        }
-      >
-        <View>
-          <Expandable
-            arrowColor={colors.greenUI}
-            containerStyle={{
-              ...styles.expandableContainer,
-              paddingVertical: isExpandable ? (expanded ? 22 : 27) : 16,
-            }}
-            isExpandable={isExpandable}
-            isExpanded={expanded}
-          >
-            {isExpandable ? (
-              <>
-                <View style={styles.left}>
-                  <Text style={styles.category}>
-                    {paymentMethod === PaymentMethod.Card
-                      ? t('selectProviderScreen.card')
-                      : t('selectProviderScreen.bank')}
-                  </Text>
-                  {!expanded && (
-                    <Text style={styles.fee}>
-                      {renderFeeAmount(sectionQuotes[0].quote, t('selectProviderScreen.minFee'))}
-                    </Text>
-                  )}
-                </View>
-
-                <View style={styles.right}>
-                  <Text style={styles.providerDropdown}>
-                    {t('selectProviderScreen.numProviders', { count: sectionQuotes.length })}
-                  </Text>
-                </View>
-              </>
-            ) : (
-              <>
-                <View style={styles.left}>
-                  <Text style={styles.category}>
-                    {paymentMethod === PaymentMethod.Card
-                      ? t('selectProviderScreen.card')
-                      : t('selectProviderScreen.bank')}
-                  </Text>
-                  <Text style={styles.fee}>
-                    {renderFeeAmount(sectionQuotes[0].quote, t('selectProviderScreen.fee'))}
-                  </Text>
-                  <Text style={styles.topInfo}>
-                    {t('selectProviderScreen.idRequired')} |{' '}
-                    {paymentMethod === PaymentMethod.Card
-                      ? t('selectProviderScreen.oneHour')
-                      : t('selectProviderScreen.numDays')}
-                  </Text>
-                </View>
-
-                <View style={styles.imageContainer}>
-                  <Image
-                    testID={`image-${sectionQuotes[0].provider.name}`}
-                    source={{ uri: sectionQuotes[0].provider.logoWide }}
-                    style={styles.providerImage}
-                    resizeMode="center"
-                  />
-                </View>
-              </>
-            )}
-          </Expandable>
-        </View>
-      </Touchable>
-      {expanded &&
-        sectionQuotes.map((cicoQuote, index) => (
-          <Touchable onPress={(quoteOnPress(cicoQuote) as unknown) as () => void}>
-            <View style={styles.expandedContainer}>
-              <View style={styles.left}>
-                <Text style={styles.expandedFee}>
-                  {renderFeeAmount(cicoQuote.quote, t('selectProviderScreen.fee'))}
-                </Text>
-                <Text style={styles.expandedInfo}>
-                  {t('selectProviderScreen.idRequired')} |{' '}
-                  {paymentMethod === PaymentMethod.Card
-                    ? t('selectProviderScreen.oneHour')
-                    : t('selectProviderScreen.numDays')}
-                </Text>
-                {index === 0 && (
-                  <Text style={styles.expandedTag}>{t('selectProviderScreen.bestRate')}</Text>
-                )}
-              </View>
-
-              <View style={styles.imageContainer}>
-                <Image
-                  testID={`image-${cicoQuote.provider.name}`}
-                  source={{ uri: cicoQuote.provider.logoWide }}
-                  style={styles.providerImage}
-                  resizeMode="center"
-                />
-              </View>
-            </View>
-          </Touchable>
-        ))}
     </View>
   )
 }
@@ -439,52 +267,12 @@ const styles = StyleSheet.create({
     flexDirection: 'column',
     justifyContent: 'center',
   },
-  expandedContainer: {
-    borderTopWidth: 1,
-    borderTopColor: colors.gray2,
-    paddingVertical: 16,
-    paddingHorizontal: 16,
-    backgroundColor: '#F9FAFA',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    flexDirection: 'row',
-  },
-  providerImage: {
-    flex: 1,
-  },
-  imageContainer: {
-    width: 80,
-    height: 40,
-  },
   category: {
     ...fontStyles.small500,
   },
   fee: {
     ...fontStyles.regular500,
     marginTop: 4,
-  },
-  providerDropdown: {
-    ...fontStyles.small500,
-    color: colors.greenUI,
-  },
-  expandedInfo: {
-    ...fontStyles.small,
-    color: colors.gray4,
-    marginTop: 2,
-  },
-  topInfo: {
-    ...fontStyles.small,
-    color: colors.gray4,
-    marginTop: 4,
-  },
-  expandedFee: {
-    ...fontStyles.regular500,
-  },
-  expandedTag: {
-    ...fontStyles.label,
-    color: colors.greenUI,
-    fontSize: 12,
-    marginTop: 2,
   },
   linkToOtherScreen: {
     ...fontStyles.small500,
